refactor: move data definitions to impl module and improve null safety
- Move `LocalSource`, `RemoteSource`, `GitHubRepositoriesDao`, `GitHubRepositoriesService`, and `NoMoreItemsException` from `data/bridge` to `data/impl` package
- Update `DataModule`, `GitHubReposRepositoryImpl`, and related classes to reflect the package structure changes
- Change return types of `LocalSource` and `GitHubRepositoriesDao` from nullable `List?` to non-nullable `List`
- Refactor `EntityMappers` to utilize Kotlin's `.orEmpty()` extension for string mapping
