+++ data/impl/src/main/java/com/caioluis/githubpopular/data/impl/di/DataModule.kt
@@ -1,16 +1,16 @@
 package com.caioluis.githubpopular.data.impl.di
 
 import android.content.Context
-import com.caioluis.githubpopular.data.bridge.local.LocalSource
-import com.caioluis.githubpopular.data.bridge.local.dao.GitHubRepositoriesDao
-import com.caioluis.githubpopular.data.bridge.remote.RemoteSource
-import com.caioluis.githubpopular.data.bridge.remote.service.GitHubRepositoriesService
 import com.caioluis.githubpopular.data.impl.BuildConfig
 import com.caioluis.githubpopular.data.impl.local.GitHubReposDataBase
 import com.caioluis.githubpopular.data.impl.local.LocalSourceImpl
+import com.caioluis.githubpopular.data.impl.local.dao.GitHubRepositoriesDao
+import com.caioluis.githubpopular.data.impl.local.model.LocalSource
 import com.caioluis.githubpopular.data.impl.remote.GitHubReposRepositoryImpl
+import com.caioluis.githubpopular.data.impl.remote.RemoteSource
 import com.caioluis.githubpopular.data.impl.remote.RemoteSourceImpl
 import com.caioluis.githubpopular.data.impl.remote.ServiceBuilder
+import com.caioluis.githubpopular.data.impl.remote.service.GitHubRepositoriesService
 import com.caioluis.githubpopular.domain.bridge.repository.GitHubReposRepository
 import dagger.Binds
 import dagger.Module
