+++ data/impl/src/main/java/com/caioluis/githubpopular/data/impl/local/LocalSourceImpl.kt
@@ -1,9 +1,9 @@
 package com.caioluis.githubpopular.data.impl.local
 
-import com.caioluis.githubpopular.data.bridge.local.LocalSource
-import com.caioluis.githubpopular.data.bridge.local.dao.GitHubRepositoriesDao
 import com.caioluis.githubpopular.data.bridge.mappers.toDomain
 import com.caioluis.githubpopular.data.bridge.mappers.toLocal
+import com.caioluis.githubpopular.data.impl.local.dao.GitHubRepositoriesDao
+import com.caioluis.githubpopular.data.impl.local.model.LocalSource
 import com.caioluis.githubpopular.domain.bridge.entity.DomainGitHubRepository
 import javax.inject.Inject
 
@@ -26,7 +26,7 @@ constructor(
     override suspend fun getFromCache(
         page: Int,
         language: String,
-    ): List<DomainGitHubRepository>? = gitHubRepositoriesDao.getAllRepositories(page, language)?.map {
+    ): List<DomainGitHubRepository> = gitHubRepositoriesDao.getAllRepositories(page, language).map {
         it.toDomain()
     }
 }
