+++ data/impl/src/main/java/com/caioluis/githubpopular/data/impl/remote/RemoteSource.kt
@@ -1,4 +1,4 @@
-package com.caioluis.githubpopular.data.bridge.remote
+package com.caioluis.githubpopular.data.impl.remote
 
 import com.caioluis.githubpopular.data.bridge.remote.model.RemoteGitHubRepository
