+++ data/impl/src/main/java/com/caioluis/githubpopular/data/impl/remote/RemoteSourceImpl.kt
@@ -1,8 +1,7 @@
 package com.caioluis.githubpopular.data.impl.remote
 
-import com.caioluis.githubpopular.data.bridge.remote.RemoteSource
 import com.caioluis.githubpopular.data.bridge.remote.model.RemoteGitHubRepository
-import com.caioluis.githubpopular.data.bridge.remote.service.GitHubRepositoriesService
+import com.caioluis.githubpopular.data.impl.remote.service.GitHubRepositoriesService
 import javax.inject.Inject
 
 class RemoteSourceImpl
